pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssae-codi-stage1"
version = "0.1.0"
description = "Stage 1 CODI model setup scaffold for inference and latent inspection."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "huggingface-hub>=0.24.0",
  "PyYAML>=6.0.2",
  "safetensors>=0.4.5",
  "torch>=2.4.0",
  "transformers>=4.44.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
